requirements.txt
pandas==1.5.3
Pillow==9.4.0
python_pptx==0.6.21
openpyxl==3.0.10